Mesh: Move Wm4 conversions into their own header
Core/Utilities.h held nothing but the Base vec_traits specialisations for
Wm4::Vector3d and Wm4::Vector3f, so the name said less than the contents did
and the file sat among the mesh core algorithms rather than with the other
conversions.

Renamed to Mesh/App/Conversion/Wm4.h, matching Gui/Conversion and
Part/App/Conversion. The content is unchanged.
